Fixes, heartbeats, duplicate detection, types, peer management, bump version

+++ components/robusto/include/robusto_peer.h
@@ -146,8 +146,8 @@ extern "C"
     robusto_peer_t *robusto_add_init_new_peer_i2c(const char *peer_name, const uint8_t i2c_address);
 #endif
 #ifdef CONFIG_ROBUSTO_SUPPORTS_CANBUS
-    robusto_peer_t *add_peer_by_canbus_address(const char *peer_name, uint32_t canbus_address);
-    robusto_peer_t *robusto_add_init_new_peer_canbus(const char *peer_name, const uint32_t canbus_address);
+    robusto_peer_t *add_peer_by_canbus_address(const char *peer_name, uint8_t canbus_address);
+    robusto_peer_t *robusto_add_init_new_peer_canbus(const char *peer_name, const uint8_t canbus_address);
 #endif
 
     robusto_peer_t *robusto_add_init_new_peer(const char *peer_name, rob_mac_address *mac_address, robusto_media_types media_types);
@@ -194,7 +194,7 @@ extern "C"
      * @param canbus_address The CAN bus address we are looking for
      * @return robusto_peer_t* 
      */
-    robusto_peer_t *robusto_peers_find_peer_by_canbus_address(uint32_t canbus_address);
+    robusto_peer_t *robusto_peers_find_peer_by_canbus_address(uint8_t canbus_address);
 #endif 
 
 
@@ -252,7 +252,7 @@ extern "C"
         uint8_t i2c_address;
 #endif
 #ifdef CONFIG_ROBUSTO_SUPPORTS_CANBUS
-        uint32_t canbus_address;
+        uint8_t canbus_address;
 #endif
     } relation_t;
 
@@ -295,7 +295,7 @@ extern "C"
 #endif
 #ifdef CONFIG_ROBUSTO_SUPPORTS_CANBUS
                       ,
-                      uint32_t canbus_address
+                      uint8_t canbus_address
 #endif
     );
